ai/requirements.txt
crewai
crewai-tools
python-dotenv
langchain-openai
pyyaml
beautifulsoup4
lxml